agent_google_adk/pyproject.toml
[project]
name = "openai-agent"
version = "1.0.0"
description = "Example of using Google's ADK with gpt-oss 20b model"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.13.0",
    "litellm>=1.76.1",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["I", "E", "F"]
ignore = ["E501"]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.rules]
non-subscriptable = "warn"
unused-ignore-comment = "warn"
invalid-argument-type = "warn"
invalid-parameter-default = "warn"
unresolved-import = "warn"
invalid-assignment = "warn"

[dependency-groups]
dev = [
    "pre-commit==4.3.0",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-mock==3.14.1",
    "ruff==0.12.10",
    "ty==0.0.1a8",
    "ipykernel==6.29.5",
]
